feat(add takes) - refactor TakesService to it's own module. Show "Add          Takes" button only when needed
